requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
email-validator==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
alembic==1.13.2
WTForms==3.1.2
pandas==2.2.2
requests==2.32.3
openpyxl==3.1.5
